# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "经济法知识图谱与上市公司治理合规推理系统 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/compliance-checks/precheck": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "Precheck Compliance Check",
        "description": "Read-only dry run — no ComplianceCheck is created, no audit event is\nrecorded, no Idempotency-Key is required. Lets the wizard show which\nrules are applicable and which facts are missing before the user\ncommits to a real, audited check.",
        "operationId": "precheck_compliance_check_api_v1_compliance_checks_precheck_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "subject_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Subject Id"
            }
          },
          {
            "name": "evaluation_time",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Evaluation Time"
            }
          },
          {
            "name": "ruleset_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ruleset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrecheckOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/compliance-checks": {
      "post": {
        "tags": [
          "compliance"
        ],
        "summary": "Create Compliance Check",
        "operationId": "create_compliance_check_api_v1_compliance_checks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplianceCheckCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceCheckOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "List Compliance Checks",
        "operationId": "list_compliance_checks_api_v1_compliance_checks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "subject_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Id"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ComplianceCheckStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ComplianceCheckOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/compliance-checks/{check_id}": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "Get Compliance Check",
        "operationId": "get_compliance_check_api_v1_compliance_checks__check_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "check_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Check Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceCheckOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/conclusions/{conclusion_id}/proof": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "Get Conclusion Proof",
        "operationId": "get_conclusion_proof_api_v1_conclusions__conclusion_id__proof_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conclusion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Conclusion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProofOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComplianceCheckCreate": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "subject_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          },
          "company_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "evaluation_time": {
            "type": "string",
            "format": "date-time",
            "title": "Evaluation Time"
          },
          "ruleset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ruleset Id"
          },
          "rule_codes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Codes"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "evaluation_time"
        ],
        "title": "ComplianceCheckCreate",
        "description": "Formal contract (06-MVP骨架充实与功能闭环计划.md §3.1): `subject_id` +\n`ruleset_id`, `Idempotency-Key` from the request header.\n\n`company_id` / `rule_codes` / body `idempotency_key` are deprecated\naliases kept for one dev cycle — the route records which were used in\nthe response's `deprecations` list."
      },
      "ComplianceCheckOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "subject_id": {
            "type": "string",
            "format": "uuid",
            "title": "Subject Id"
          },
          "evaluation_time": {
            "type": "string",
            "format": "date-time",
            "title": "Evaluation Time"
          },
          "rule_set_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Set Id"
          },
          "ruleset_snapshot": {
            "items": {},
            "type": "array",
            "title": "Ruleset Snapshot"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "conclusions": {
            "items": {
              "$ref": "#/components/schemas/ConclusionOut"
            },
            "type": "array",
            "title": "Conclusions"
          },
          "deprecations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Deprecations",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "subject_id",
          "evaluation_time",
          "rule_set_id",
          "ruleset_snapshot",
          "status",
          "conclusions"
        ],
        "title": "ComplianceCheckOut"
      },
      "ComplianceCheckStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "COMPLETED",
          "FAILED"
        ],
        "title": "ComplianceCheckStatus"
      },
...
      "ConclusionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "rule_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Rule Version Id"
          },
          "rule_code": {
            "type": "string",
            "title": "Rule Code"
          },
          "rule_name": {
            "type": "string",
            "title": "Rule Name"
          },
          "result_status": {
            "type": "string",
            "title": "Result Status"
          },
          "missing_facts": {
            "items": {},
            "type": "array",
            "title": "Missing Facts"
          },
          "applicable_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicable Reason"
          },
          "excluded_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excluded Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "rule_version_id",
          "rule_code",
          "rule_name",
          "result_status",
          "missing_facts",
          "applicable_reason",
          "excluded_reason"
        ],
        "title": "ConclusionOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_ComplianceCheckOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ComplianceCheckOut"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "Page[ComplianceCheckOut]"
      },
...
      "PrecheckItemOut": {
        "properties": {
          "rule_code": {
            "type": "string",
            "title": "Rule Code"
          },
          "rule_name": {
            "type": "string",
            "title": "Rule Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "missing_facts": {
            "items": {},
            "type": "array",
            "title": "Missing Facts"
          },
          "applicable_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicable Reason"
          },
          "excluded_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excluded Reason"
          }
        },
        "type": "object",
        "required": [
          "rule_code",
          "rule_name",
          "status",
          "missing_facts",
          "applicable_reason",
          "excluded_reason"
        ],
        "title": "PrecheckItemOut"
      },
      "PrecheckOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PrecheckItemOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PrecheckOut"
      },
      "ProofOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "conclusion_id": {
            "type": "string",
            "format": "uuid",
            "title": "Conclusion Id"
          },
          "root_step_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Step Id"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/ProofStepOut"
            },
            "type": "array",
            "title": "Steps"
          }
        },
        "type": "object",
        "required": [
          "id",
          "conclusion_id",
          "root_step_id",
          "steps"
        ],
        "title": "ProofOut"
      },
      "ProofStepOut": {
        "properties": {
          "sequence_no": {
            "type": "integer",
            "title": "Sequence No"
          },
          "step_type": {
            "type": "string",
            "title": "Step Type"
          },
          "rule_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Version Id"
          },
          "rule_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Id"
          },
          "rule_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Code"
          },
          "input_facts": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Facts"
          },
          "calculation": {
            "additionalProperties": true,
            "type": "object",
            "title": "Calculation"
          },
          "output_state": {
            "additionalProperties": true,
            "type": "object",
            "title": "Output State"
          }
        },
        "type": "object",
        "required": [
          "sequence_no",
          "step_type",
          "rule_version_id",
          "rule_id",
          "rule_code",
          "input_facts",
          "calculation",
          "output_state"
        ],
        "title": "ProofStepOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}